Move persist_builtin_flow from FlowRegistryExt to ConnectedFlowRegistry
